change controller url
