+++ src/main/java/lab/reservation_server/controller/LectureController.java
@@ -8,12 +8,14 @@ import lab.reservation_server.dto.request.lecture.LectureEditDto;
 import lab.reservation_server.dto.request.lecture.LectureSaveDto;
 import lab.reservation_server.dto.response.DefaultMessageResponse;
 import lab.reservation_server.dto.response.lecture.LectureInfo;
+import lab.reservation_server.service.LabService;
 import lab.reservation_server.service.LectureService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -23,11 +25,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @Validated
 @RequiredArgsConstructor
-@Api(tags = "Lecture Controller : 실습실 강의 시간표 관련")
+@Api(tags = "Lecture Controller : 실습실 강의 시간표 관련, 시간표 추가 수정 삭제")
 @Slf4j
 public class LectureController {
 
     private final LectureService lectureService;
+    private final LabService labService;
 
     /**
      * 강의 시간표 추가
@@ -71,6 +74,16 @@ public class LectureController {
         return ResponseEntity.ok(new DefaultMessageResponse("강의 시간표 삭제 성공"));
     }
 
+
+    /**
+     * 강의실 전체에 대한 모든 시간표 조회
+     */
+    @GetMapping("/api/lectures")
+    @ApiOperation(value="모든 강의실 전체 시간표 조회" , notes = "강의실 전체에 대한 시간표를 조회할 수 있다.")
+    public ResponseEntity<List<LectureInfo>> getAllLabTimeTable() {
+        List<LectureInfo> allTimeTable = labService.getAllLabTimeTable();
+        return ResponseEntity.ok(allTimeTable);
+    }
